+++ TinkoffASDKCore/TinkoffASDKCoreTests/Infrastructure/Mocks/LanguageProviderMock.swift
@@ -1,0 +1,13 @@
+//
+//  LanguageProviderMock.swift
+//  TinkoffASDKCore-Unit-Tests
+//
+//  Created by Никита Васильев on 21.07.2023.
+//
+
+import Foundation
+@testable import TinkoffASDKCore
+
+final class LanguageProviderMock: ILanguageProvider {
+    var language: AcquiringSdkLanguage?
+}