+++ CHANGELOG.md
@@ -37,6 +37,10 @@
 * [EACQAPW-5581] CardDataFormatter Tests
 * [EACQAPW-5583] URLDataLoaderTests Tests
 * [EACQAPW-5597] ThreeDSURLRequestBuilder Tests
+* [EACQAPW-5586] Added to Sample Option to choose App Based Sdk Interface
+* [EACQAPW-5597] ThreeDSURLBuilder Tests
+* [EACQAPW-5606] IPAddressFactory Tests
+* [EACQAPW-5607] ThreeDSDeviceInfoProvider Tests
 
 ### Fixed